Individual computes fitness at construction The individual class now computes fitness when first constructed. - all contructors for individual now takes domain as at least one of the paraemeters - also fixed compute fitness so it does float division instead of integer division

+++ src/Evolutionary/Domain.java
@@ -18,7 +18,7 @@ class Domain extends AbstractDomain {
     public double computeFitness(Individual i, Domain d){
         String gen = i.getGenMak();
         int ones = gen.length() - gen.replaceAll("1", "").length();
-        return ones / d.getBitLength();
+        return (float)ones / (float)d.getBitLength();
 
     }
 }
